+++ db/sqlc/entry.sql.go
@@ -18,8 +18,8 @@ returning id, account_id, amount, created_at
 `
 
 type CreateEntryParams struct {
-	AccountID int64   `json:"account_id"`
-	Amount    float64 `json:"amount"`
+	AccountID int64 `json:"account_id"`
+	Amount    int64 `json:"amount"`
 }
 
 func (q *Queries) CreateEntry(ctx context.Context, arg CreateEntryParams) (Entry, error) {
